requirements.txt
dataclasses~=0.6
kubernetes~=12.0.1
sqlalchemy < 1.4.0
pandas~=1.2.3
PyYAML~=5.4.1

pathlib~=1.0.1
requests~=2.25.1
urllib3~=1.26.4
apache-airflow-providers-amazon